[Bronze III] Title: Simple Arithmetic, Time: 60 ms, Memory: 36456 KB, Score: 100 point -BaekjoonHub

+++ Python/백준/Bronze/15858. Simple Arithmetic/Simple Arithmetic.py
@@ -1,2 +1,3 @@
-a, b, c = map(int, input().split())
+from decimal import Decimal
+a, b, c = map(Decimal, input().split())
 print(a*b/c)
